Avoid spurious 'Too few positionals...' on regex
If the backtraces are to be believed, Raku seemingly calls the
parametrised `kw` regex with no arguments sometimes. (Perhaps during
backtracking?)

Turning the parameter optional seems to prevent those 'Too few
positionals passed' errors, even though I never seem to be able to catch
Raku in the act of calling the method with no argument!